Сделал преобразование структуры проекта
