+++ internal/app/router/router.go
@@ -4,10 +4,10 @@ import (
 	"log"
 	"runtime/debug"
 
-	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api"
 	comm "github.com/denlipov/omp-bot/internal/app/commands/communication"
+	"github.com/denlipov/omp-bot/internal/app/commands/demo"
 	"github.com/denlipov/omp-bot/internal/app/path"
-        "github.com/denlipov/omp-bot/internal/app/commands/demo"
+	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api"
 )
 
 type Commander interface {
@@ -21,7 +21,7 @@ type Router struct {
 
 	// demoCommander
 	demoCommander Commander
-        commCommander Commander
+	commCommander Commander
 	// user
 	// access
 	// buy
@@ -57,7 +57,7 @@ func NewRouter(
 		bot: bot,
 		// demoCommander
 		demoCommander: demo.NewDemoCommander(bot),
-                commCommander: comm.NewCommunicationCommander(bot),
+		commCommander: comm.NewCommunicationCommander(bot),
 		// user
 		// access
 		// buy
@@ -111,8 +111,8 @@ func (c *Router) handleCallback(callback *tgbotapi.CallbackQuery) {
 	switch callbackPath.Domain {
 	case "demo":
 		c.demoCommander.HandleCallback(callback, callbackPath)
-        case "communication":
-                c.commCommander.HandleCallback(callback, callbackPath)
+	case "communication":
+		c.commCommander.HandleCallback(callback, callbackPath)
 	case "user":
 		break
 	case "access":
@@ -184,8 +184,8 @@ func (c *Router) handleMessage(msg *tgbotapi.Message) {
 	switch commandPath.Domain {
 	case "demo":
 		c.demoCommander.HandleCommand(msg, commandPath)
-        case "communication":
-                c.commCommander.HandleCommand(msg, commandPath)
+	case "communication":
+		c.commCommander.HandleCommand(msg, commandPath)
 	case "user":
 		break
 	case "access":
